% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{match_names}
\alias{match_names}
\title{Package for searching names of plants and their synonyms mentioned in Wildlife (Protection) Act, 1972 (Amended as on date)'}
\usage{
match_names()
}
\description{
Package for searching names of plants and their synonyms mentioned in Wildlife (Protection) Act, 1972 (Amended as on date)'
}
